% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Buss1989}
\alias{Buss1989}
\title{Buss1989}
\format{
A data frame with 148 rows and 6 variables:
\describe{
  \item{\code{Sample}}{integer Sample population}
  \item{\code{Region}}{integer World region for sample}
  \item{\code{question}}{character Which study item}
  \item{\code{Mean_F}}{double Mean value for female respondents}
  \item{\code{Mean_M}}{double Mean value for male respondents}
  \item{\code{SD_F}}{double Standard deviation for female respondents}
  \item{\code{SD_M}}{double Standard deviation for male respondents}
}
}
\usage{
Buss1989
}
\description{
Summary statistics from David Buss's 1989 study "Sex differences in human mate preferences: Evolutionary hypotheses tested in 37 cultures". Caution: these data were scraped directly from an old pdf and have not been vetted for errors!
}
\keyword{datasets}
