% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConformistTransmission.R
\name{ConformistTransmission}
\alias{ConformistTransmission}
\title{ConformistTransmission}
\usage{
ConformistTransmission(N, p_0, D, t_max, add = FALSE, col = "black")
}
\arguments{
\item{N}{Population size}

\item{p_0}{Proportion of agents with trait A at the start of the simulation.}

\item{D}{Strength of conformity}

\item{t_max}{How many generations to run the sim.}

\item{add}{Whether or not the new sim should be added to an existing plot; defaults to FALSE.}

\item{col}{Color of lines for plotting; defaults to black.}
}
\description{
Agent-based model simulation of conformist tranmission, only slightly modified from Alex Mesoudi's code here: https://github.com/amesoudi/cultural_evolution_ABM_tutorial/blob/master/ABMmodels_model3.Rmd
}
