% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HipsterTransmission.R
\name{HipsterTransmission}
\alias{HipsterTransmission}
\title{HipsterTransmission}
\usage{
HipsterTransmission(N, d, s = 0, k, t_max)
}
\arguments{
\item{N}{Population size}

\item{d}{Anticonformity preference, units indicate standard deviations. 0 indicates preference for mean, positive values prefer higher than mean, negative values prefer lower.}

\item{s}{Variation in anticonformity preference, units indicate standard deviations; defaults to 0.}

\item{k}{Adjustment rate, bounded between 0 and 1. Controls how fast agents update their trait values}

\item{t_max}{How many generations to run the sim.}
}
\description{
Agent-based model simulation of anticorformity in a quantitative trait, based on Smaldino and Epstein 'Social conformity despite individual preference for distinctiveness'
}
