% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{makeTransparent}
\usage{
makeTransparent(..., alpha = 0.5)
}
\arguments{
\item{alpha}{How transparent should the color be? Between 0 and 1}
}
\description{
Convienience function for applying transparency to a color, taken from Ricardo Oliveros-Ramos' stackoverflow post: https://stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color
}
