% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnbiasedTransmission.R
\name{UnbiasedTransmission}
\alias{UnbiasedTransmission}
\title{UnbiasedTransmission}
\usage{
UnbiasedTransmission(N, t_max, add = FALSE, col = "black")
}
\arguments{
\item{N}{Population size}

\item{t_max}{How many generations to run the sim.}

\item{add}{Whether or not the new sim should be added to an existing plot; defaults to FALSE.}

\item{col}{Color of lines for plotting; defaults to black.}
}
\description{
Agent-based model simulation of unbiased tranmission, only slightly modified from Alex Mesoudi's code here: https://github.com/amesoudi/cultural_evolution_ABM_tutorial/blob/master/ABMmodels_model1.Rmd
}
