% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peppered_moths.R
\name{moth_sim}
\alias{moth_sim}
\title{moth_sim}
\usage{
moth_sim(pop_size, pollution, speed, predation, drift)
}
\arguments{
\item{pop_size}{Initial moth population size}

\item{pollution}{Environmental pollution levels, MUST BE BETWEEN 0 and 1.}

\item{speed}{Rate of environmental (pollution) change, either 'fixed' (no change), 'slow', or 'fast'.}

\item{predation}{Intensity of predation, MUST BE BETWEEN 0 and 1.}

\item{drift}{magnitude of drift, must be either 'none', 'small', or 'large'.}
}
\value{
Returns a list of moth phenotypes over time, plus pollution levels and input parameter values
}
\description{
Simulates the evolution of peppered moths in a polluted environment, based on Wilensky, U. (1997). NetLogo Peppered Moths model. http://ccl.northwestern.edu/netlogo/models/PepperedMoths. Center for Connected Learning and Computer-Based Modeling, Northwestern University, Evanston, IL.
}
