% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peppered_moths.R
\name{moth_sim}
\alias{moth_sim}
\title{moth_sim}
\usage{
moth_sim(pollution, speed, predation, mutation_rate)
}
\arguments{
\item{pollution}{Environmental pollution levels, MUST BE BETWEEN 0 and 1.}

\item{speed}{Rate of environmental (pollution) change, either 'fixed' (no change), 'slow', or 'fast'.}

\item{predation}{Intensity of predation, MUST BE BETWEEN 0 and 1.}

\item{mutation_rate}{mutation_rate, MUST BE BETWEEN 0 and 1.}
}
\value{
Returns a dataframe of moth phenotypes over time, plus pollution levels
}
\description{
Simulates the evolution of peppered moths in a polluted environment, based on Wilensky, U. (1997). NetLogo Peppered Moths model. http://ccl.northwestern.edu/netlogo/models/PepperedMoths. Center for Connected Learning and Computer-Based Modeling, Northwestern University, Evanston, IL.
}
