% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peppered_moths.R
\name{plot_moth}
\alias{plot_moth}
\title{plot_moth}
\usage{
plot_moth(sim)
}
\arguments{
\item{sim}{moth simulation data, created by moth_sim()}
}
\value{
three-panel plot
}
\description{
Plots the results of peppered moths simulation
}
